chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/import-style-tree-order-dedup-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
PASS colorOf(target1) is 'rgb(0, 128, 0)'
PASS colorOf(target2) is 'rgb(255, 0, 0)'
PASS colorOf(target3) is 'rgb(255, 0, 0)'
PASS colorOf(target1) is 'rgb(0, 128, 0)'
FAIL colorOf(target2) should be rgb(0, 0, 255). Was rgb(255, 0, 0).
PASS colorOf(target3) is 'rgb(255, 0, 0)'
PASS colorOf(target1) is 'rgb(0, 128, 0)'
FAIL colorOf(target2) should be rgb(0, 0, 255). Was rgb(255, 0, 0).
PASS colorOf(target3) is 'rgb(255, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE

Hello! Hello! Hello!
